chore: remove project modules context provider and use a hook instead
