+++ shared/src/containers/ProjectTreeTable/context/ProjectTableContext.tsx
@@ -48,8 +48,8 @@ import useBuildGroupByTableData, {
   ROW_ID_SEPARATOR,
 } from '../hooks/useBuildGroupByTableData'
 import { PowerpackContextType } from '@shared/context'
-import { ProjectTableModuleContextType } from './ProjectTableModuleContext'
 import { useColumnSettingsContext } from './ColumnSettingsContext'
+import { ProjectTableModulesType } from '../hooks'
 
 export const parseRowId = (rowId: string) => rowId.split(ROW_ID_SEPARATOR)[0] || rowId
 
@@ -121,7 +121,7 @@ export interface ProjectTableProviderProps {
   powerpack?: PowerpackContextType
 
   // remote modules
-  modules: ProjectTableModuleContextType
+  modules: ProjectTableModulesType
 
   groupByConfig?: {
     entityType?: GroupByEntityType
