--- shared/src/containers/ProjectTreeTable/context/ProjectTableModuleContext.tsx
@@ -1,90 +1,0 @@
-import { useLoadModule } from '@shared/hooks'
-import React, { createContext, useContext, ReactNode, useMemo } from 'react'
-import { GroupSettingsFallback } from '../components/GroupSettingsFallback'
-import { EntityGroup } from '@shared/api'
-import { Filter } from '@ynput/ayon-react-components'
-import { TableGroupBy } from './ColumnSettingsContext'
-import { ProjectTableAttribute } from '../types'
-
-type GetGroupQueriesParams = {
-  taskGroups: EntityGroup[]
-  filters: Filter[]
-  groupBy: TableGroupBy
-  groupPageCounts: Record<string, number>
-  dataType: ProjectTableAttribute['data']['type']
-}
-
-type GetGroupQueriesReturn = {
-  value: any
-  count: number
-  filter: string
-}[]
-
-const getGroupQueriesFallback = (params: GetGroupQueriesParams): GetGroupQueriesReturn => []
-
-export interface ProjectTableModuleContextType {
-  GroupSettings: typeof GroupSettingsFallback
-  getGroupQueries?: (params: GetGroupQueriesParams) => GetGroupQueriesReturn
-  requiredVersion?: string
-  isLoading: boolean
-}
-
-const ProjectTableModuleContext = createContext<ProjectTableModuleContextType | undefined>(
-  undefined,
-)
-
-export interface ProjectTableModuleProviderProps {
-  children: ReactNode
-  GroupSettings?: typeof GroupSettingsFallback
-  getGroupQueries?: (params: GetGroupQueriesParams) => GetGroupQueriesReturn
-}
-
-export const ProjectTableModuleProvider: React.FC<ProjectTableModuleProviderProps> = ({
-  children,
-  GroupSettings: GroupSettingsProp,
-  getGroupQueries: getGroupQueriesProp,
-}) => {
-  const minVersion = '1.0.6-dev'
-  const [GroupSettings, { outdated, isLoading: isLoadingSettings }] = useLoadModule({
-    addon: 'powerpack',
-    remote: 'slicer',
-    module: 'GroupSettings',
-    fallback: GroupSettingsProp || GroupSettingsFallback,
-    minVersion: minVersion,
-    skip: !!GroupSettingsProp,
-  })
-
-  const [getGroupQueries, { isLoading: isLoadingQueries }] = useLoadModule({
-    addon: 'powerpack',
-    remote: 'slicer',
-    module: 'getGroupQueries',
-    fallback: getGroupQueriesProp || getGroupQueriesFallback,
-    minVersion: minVersion,
-    skip: !!getGroupQueriesProp,
-  })
-
-  const isLoading = isLoadingSettings || isLoadingQueries
-  const value = useMemo(
-    () => ({
-      GroupSettings,
-      getGroupQueries,
-      requiredVersion: outdated?.required,
-      isLoading: isLoading,
-    }),
-    [GroupSettings, getGroupQueries, outdated?.required, isLoading],
-  )
-
-  return (
-    <ProjectTableModuleContext.Provider value={value}>
-      {children}
-    </ProjectTableModuleContext.Provider>
-  )
-}
-
-export const useProjectTableModuleContext = (): ProjectTableModuleContextType => {
-  const context = useContext(ProjectTableModuleContext)
-  if (context === undefined) {
-    throw new Error('useProjectTableModuleContext must be used within a ProjectTableModuleProvider')
-  }
-  return context
-}
