+++ shared/src/containers/ProjectTreeTable/context/index.ts
@@ -9,4 +9,3 @@ export * from './SelectedRowsContext'
 export * from './SelectionCellsContext'
 export * from './ProjectTableSelectionContext'
 export * from './ProjectDataContext'
-export * from './ProjectTableModuleContext'
