+++ shared/src/containers/ProjectTreeTable/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useScopedAttributeFields'
 export * from './useExpandedState'
 export * from './useColumnSorting'
 export * from './useBuildGroupByTableData'
+export * from './useProjectTableModules'
