+++ src/pages/ProjectOverviewPage/providers/ProjectOverviewDataProvider.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import {
   ColumnSettingsProvider,
   ProjectDataProvider,
-  useProjectTableModuleContext,
+  useProjectTableModules,
 } from '@shared/containers/ProjectTreeTable'
 import { SettingsPanelProvider } from '@shared/context'
 import { useAppSelector } from '@state/store'
@@ -13,17 +13,18 @@ import ProjectOverviewTableProvider from './ProjectOverviewTableProvider'
 const ProjectOverviewDataProvider: FC = () => {
   const projectName = useAppSelector((state) => state.project.name) || ''
 
-  const modules = useProjectTableModuleContext()
   const [pageConfig, updatePageConfig] = useUserProjectConfig({
     selectors: ['overview', projectName],
   })
+
+  const modules = useProjectTableModules()
 
   return (
     <ProjectDataProvider projectName={projectName}>
       <ColumnSettingsProvider config={pageConfig} onChange={updatePageConfig}>
         <ProjectOverviewProvider modules={modules}>
           <SettingsPanelProvider>
-            <ProjectOverviewTableProvider />
+            <ProjectOverviewTableProvider modules={modules} />
           </SettingsPanelProvider>
         </ProjectOverviewProvider>
       </ColumnSettingsProvider>
